Make DDIN sandbox routing dynamic and reconcile status with DDIN's real API
- Per-integrator sandbox_uses_real_provider flag routes only the internal
  test integrator's sandbox key to real DDIN, keeping every other
  integrator's sandbox safely simulated by default.
- Add a terminal DEBIT_FAILED status so a failed Fineract debit no longer
  gets stuck behind a misleading 409 on retry.
- Poll DDIN's real GET /v1/momo/collection/{referenceId} to reconcile a
  DEBITED transaction the moment DDIN resolves it, instead of only relying
  on a webhook that can't reach localhost: auto-sync on view, an explicit
  sync endpoint, and a 30s background reconciliation loop.
- Fix provider success-rate math to exclude in-flight transactions instead
  of counting them as failures.
- Replace internal bookkeeping labels (DEBITED, DEBIT_FAILED) with DDIN's
  real Pending/Success/Failed vocabulary everywhere a transaction status is
  shown to a user; show the real customer name in the dashboard table.
- Fix a hardcoded "Live mock mode" label that never reflected the actual
  mock-mode state; pin all timestamps to Africa/Kigali (CAT).
- Add fast client-side polling on an in-flight collection result so it
  resolves on screen within seconds instead of waiting on the background
  sweep.

+++ app/schemas/collection.py
@@ -1,4 +1,3 @@
-from decimal import Decimal
 from typing import Literal, Optional
 
 from pydantic import BaseModel, Field, condecimal
@@ -28,6 +27,11 @@ class CollectionResponse(BaseModel):
     debit_transaction_id: Optional[str] = None
     refund_transaction_id: Optional[str] = None
     provider_transaction_reference: Optional[str] = None
-    amount_rwf: Decimal
+    # float, not Decimal: this is the response/output side - Pydantic v2
+    # serializes Decimal to a JSON *string* (to preserve precision), which
+    # silently breaks frontend code expecting a number. CollectionRequest
+    # above (the input side) correctly keeps condecimal for validation; the
+    # orchestrator's internal math is still all Decimal.
+    amount_rwf: float
     message: str
     refunded: bool = False
